add ft_memcpy

+++ Libft/ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: hluiz-ma <hluiz-ma@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/14 12:03:18 by hluiz-ma          #+#    #+#             */
-/*   Updated: 2024/04/14 19:58:45 by hluiz-ma         ###   ########.fr       */
+/*   Updated: 2024/04/21 18:38:02 by hluiz-ma         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@
 
 void	*ft_memcpy(void *dst, const void *src, size_t n)
 {
-    size_t i;
+    unsigned int i;
 
     
     if(!dst && !src)
